Resolve substitutions only after all files have been loaded.
This allows included files to properly reference keys that are defined
in an including file. From the HOCON spec:

    Substitution processing is performed as the last parsing step, so a
    substitution can look forward in the configuration. If a
    configuration consists of multiple files, it may even end up
    retrieving a value from another file.

This change has the possibility of breaking configuration files. For
example:

foo.conf:

    { x: 10, y: ${x} }

bar.conf:

    include "foo.conf"
    x: 42

Prior to this change, the output of pyhocon is:

    {
      "x": 42,
      "y": 10
    }

because the value of `y` is looked up immediately after `foo.conf` is
parsed.

After this change, the output of pyhocon is:

    {
      "x": 42,
      "y": 42
    }

This matches what the reference java implementation produces.

This change does not address "fixing up" as described in
https://github.com/typesafehub/config/blob/master/HOCON.md#include-semantics-substitution