Fixed Python 2.6 incompatible format string
pyhocon officially supports Python 2.6, so it can't use the {} syntax in format strings ([which was added in 2.7](https://docs.python.org/2/library/string.html#format-string-syntax)). This is the only instance in the code.

Admittedly, I don't know what input would hit this case, so I cannot create a test case for it (It is one of the few lines not currently in the code coverage).